SN1-152: Fixing miner by downgrading bittensor version
