+++ .env.example
@@ -1,4 +1,3 @@
-TEST=true # whether you are running in test mode
 NET_UID=61 # the network UID. If 1, we automatically assume you're running on the mainnet, for any other value, we assume you're running on the testnet (with test settings)
 WALLET_NAME="validator" # the name of your wallet
 HOTKEY="validator_hotkey" # the name of the hotkey associated with the validator wallet
