+++ prompting/base/miner.py
@@ -1,48 +1,60 @@
 import time
-import asyncio
+import wandb
+import prompting
 import threading
 import bittensor as bt
+from datetime import datetime
 from prompting.base.protocol import StreamPromptingSynapse
 from prompting.base.neuron import BaseNeuron
 from traceback import print_exception
 from prompting import settings
+from loguru import logger
+from pydantic import BaseModel, model_validator, ConfigDict
+from typing import Tuple
 
 
-class BaseStreamMinerNeuron(BaseNeuron):
+class BaseStreamMinerNeuron(BaseModel, BaseNeuron):
     """
     Base class for Bittensor miners.
     """
 
-    def __init__(self, config=None):
-        super().__init__(config=config)
+    step: int = 0
+    wallet: bt.wallet | None = None
+    axon: bt.axon | None = None
+    subtensor: bt.subtensor | None = None
+    metagraph: bt.metagraph | None = None
+    should_exit: bool = False
+    is_running: bool = False
+    thread: threading.Thread = None
+    uid: int | None = None
 
-        # Warn if allowing incoming requests from anyone.
-        if not settings.BLACKLIST_FORCE_VALIDATOR_PERMIT:
-            bt.logging.warning(
-                "You are allowing non-validators to send requests to your miner. This is a security risk."
-            )
-        if settings.BLACKLIST_ALLOW_NON_REGISTERED:
-            bt.logging.warning(
-                "You are allowing non-registered entities to send requests to your miner. This is a security risk."
-            )
+    model_config = ConfigDict(arbitrary_types_allowed=True)
 
-        # The axon handles request processing, allowing validators to send this miner requests.
-        self.axon = bt.axon(wallet=self.wallet, config=self.config)
-
-        # Attach determiners which functions are called when servicing a request.
-        bt.logging.info("Attaching forward function to miner axon.")
+    @model_validator(mode="after")
+    def attach_axon(self) -> "BaseStreamMinerNeuron":
+        # note that this initialization has to happen in the validator because the objects
+        # are not picklable and because pydantic deepcopies things it breaks
+        self.wallet = bt.wallet(name=settings.WALLET_NAME, hotkey=settings.HOTKEY)
+        self.axon = bt.axon(wallet=self.wallet, port=settings.AXON_PORT)
+        logger.info("Attaching axon")
         self.axon.attach(
             forward_fn=self._forward,
             blacklist_fn=self.blacklist,
             priority_fn=self.priority,
         )
-        bt.logging.info(f"Axon created: {self.axon}")
+        # self.axon.attach(
+        #     forward_fn=self.test_log,
+        #     blacklist_fn=self.blacklist,
+        #     priority_fn=self.priority,
+        # )
+        self.subtensor = bt.subtensor(network=settings.SUBTENSOR_NETWORK)
+        self.metagraph = bt.metagraph(netuid=settings.NETUID, network=settings.SUBTENSOR_NETWORK, sync=True, lite=False)
+        self.uid = self.metagraph.hotkeys.index(self.wallet.hotkey.ss58_address)
+        logger.info(f"Axon created: {self.axon}; running on uid: {self.uid}")
+        self.axon.serve(netuid=settings.NETUID, subtensor=self.subtensor)
+        return self
 
-        # Instantiate runners
-        self.should_exit: bool = False
-        self.is_running: bool = False
-        self.thread: threading.Thread = None
-        self.lock = asyncio.Lock()
+        # self.lock = asyncio.Lock()
 
     def run(self):
         """
@@ -72,18 +84,17 @@ class BaseStreamMinerNeuron(BaseNeuron):
 
         # Serve passes the axon information to the network + netuid we are hosting on.
         # This will auto-update if the axon port of external ip have changed.
-        bt.logging.info(f"Serving miner axon {self.axon} with netuid: {settings.NETUID}")
-        self.axon.serve(netuid=settings.NETUID, subtensor=self.subtensor)
+        logger.info(f"Serving miner axon {self.axon} with netuid: {settings.NETUID}")
 
         # Start  starts the miner's axon, making it active on the network.
         self.axon.start()
 
-        bt.logging.info(f"Miner starting at block: {self.block}")
+        logger.info(f"Miner starting at block: {self.subtensor.get_current_block()}")
         last_update_block = 0
         # This loop maintains the miner's operations until intentionally stopped.
         try:
             while not self.should_exit:
-                while self.block - last_update_block < self.config.neuron.epoch_length:
+                while self.subtensor.get_current_block() - last_update_block < settings.NEURON_EPOCH_LENGTH:
                     # Wait before checking again.
                     time.sleep(1)
 
@@ -93,19 +104,19 @@ class BaseStreamMinerNeuron(BaseNeuron):
 
                 # Sync metagraph and potentially set weights.
                 self.sync()
-                last_update_block = self.block
+                last_update_block = self.subtensor.get_current_block()
                 self.step += 1
 
         # If someone intentionally stops the miner, it'll safely terminate operations.
         except KeyboardInterrupt:
             self.axon.stop()
-            bt.logging.success("Miner killed by keyboard interrupt.")
+            logger.success("Miner killed by keyboard interrupt.")
             exit()
 
         # In case of unforeseen errors, the miner will log the error and continue operations.
         except Exception as err:
-            bt.logging.error("Error during mining", str(err))
-            bt.logging.debug(print_exception(type(err), err, err.__traceback__))
+            logger.error("Error during mining", str(err))
+            logger.debug(print_exception(type(err), err, err.__traceback__))
             self.should_exit = True
 
     def run_in_background_thread(self):
@@ -114,23 +125,23 @@ class BaseStreamMinerNeuron(BaseNeuron):
         This is useful for non-blocking operations.
         """
         if not self.is_running:
-            bt.logging.debug("Starting miner in background thread.")
+            logger.debug("Starting miner in background thread.")
             self.should_exit = False
             self.thread = threading.Thread(target=self.run, daemon=True)
             self.thread.start()
             self.is_running = True
-            bt.logging.debug("Started")
+            logger.debug("Started")
 
     def stop_run_thread(self):
         """
         Stops the miner's operations that are running in the background thread.
         """
         if self.is_running:
-            bt.logging.debug("Stopping miner in background thread.")
+            logger.debug("Stopping miner in background thread.")
             self.should_exit = True
             self.thread.join(5)
             self.is_running = False
-            bt.logging.debug("Stopped")
+            logger.debug("Stopped")
 
     def __enter__(self):
         """
@@ -158,7 +169,7 @@ class BaseStreamMinerNeuron(BaseNeuron):
 
     def resync_metagraph(self):
         """Resyncs the metagraph and updates the hotkeys and moving averages based on the new metagraph."""
-        bt.logging.info("resync_metagraph()")
+        logger.info("resync_metagraph()")
 
         # Sync the metagraph.
         self.metagraph.sync(subtensor=self.subtensor)
@@ -186,4 +197,151 @@ class BaseStreamMinerNeuron(BaseNeuron):
             This method is not meant to be called directly but is invoked internally when a request
             is received, and it subsequently calls the `forward` method of the subclass.
         """
+        self.step += 1
+        logger.info("Calling self._forward in BaseStreamMinerNeuron")
         return self.forward(synapse=synapse)
+
+    async def blacklist(self, synapse: StreamPromptingSynapse) -> Tuple[bool, str]:
+        # WARNING: The typehint must remain Tuple[bool, str] to avoid runtime errors. YOU
+        # CANNOT change to tuple[bool, str]!!!
+        """
+        Determines whether an incoming request should be blacklisted and thus ignored. Your implementation should
+        define the logic for blacklisting requests based on your needs and desired security parameters.
+
+        Blacklist runs before the synapse data has been deserialized (i.e. before synapse.data is available).
+        The synapse is instead contructed via the headers of the request. It is important to blacklist
+        requests before they are deserialized to avoid wasting resources on requests that will be ignored.
+
+        Args:
+            synapse (StreamPromptingSynapse): A synapse object constructed from the headers of the incoming request.
+
+        Returns:
+            Tuple[bool, str]: A tuple containing a boolean indicating whether the synapse's hotkey is blacklisted,
+                            and a string providing the reason for the decision.
+
+        This function is a security measure to prevent resource wastage on undesired requests. It should be enhanced
+        to include checks against the metagraph for entity registration, validator status, and sufficient stake
+        before deserialization of synapse data to minimize processing overhead.
+
+        Example blacklist logic:
+        - Reject if the hotkey is not a registered entity within the metagraph.
+        - Consider blacklisting entities that are not validators or have insufficient stake.
+
+        In practice it would be wise to blacklist requests from entities that are not validators, or do not have
+        enough stake. This can be checked via metagraph.S and metagraph.validator_permit. You can always attain
+        the uid of the sender via a metagraph.hotkeys.index( synapse.dendrite.hotkey ) call.
+
+        Otherwise, allow the request to be processed further.
+        """
+        if synapse.dendrite.hotkey not in self.metagraph.hotkeys:
+            # Ignore requests from unrecognized entities.
+            logger.trace(f"Blacklisting unrecognized hotkey {synapse.dendrite.hotkey}")
+            return True, "Unrecognized hotkey"
+
+        logger.trace(f"Not Blacklisting recognized hotkey {synapse.dendrite.hotkey}")
+        return False, "Hotkey recognized!"
+
+    async def priority(self, synapse: StreamPromptingSynapse) -> float:
+        """
+        The priority function determines the order in which requests are handled. More valuable or higher-priority
+        requests are processed before others. You should design your own priority mechanism with care.
+
+        This implementation assigns priority to incoming requests based on the calling entity's stake in the metagraph.
+
+        Args:
+            synapse (StreamPromptingSynapse): The synapse object that contains metadata about the incoming request.
+
+        Returns:
+            float: A priority score derived from the stake of the calling entity.
+
+        Miners may recieve messages from multiple entities at once. This function determines which request should be
+        processed first. Higher values indicate that the request should be processed first. Lower values indicate
+        that the request should be processed later.
+
+        Example priority logic:
+        - A higher stake results in a higher priority value.
+        """
+        caller_uid = self.metagraph.hotkeys.index(synapse.dendrite.hotkey)  # Get the caller index.
+        priority = float(self.metagraph.S[caller_uid])  # Return the stake as the priority.
+        logger.trace(f"Prioritizing {synapse.dendrite.hotkey} with value: ", priority)
+        # priority = 1.0
+        return priority
+
+    def init_wandb(self):
+        logger.info("Initializing wandb...")
+
+        uid = f"uid_{self.metagraph.hotkeys.index(self.wallet.hotkey.ss58_address)}"
+        net_uid = f"netuid_{settings.NETUID}"
+        tags = [
+            self.wallet.hotkey.ss58_address,
+            net_uid,
+            f"uid_{uid}",
+            prompting.__version__,
+            str(prompting.__spec_version__),
+        ]
+
+        # Add uid, netuid and timestamp to run name
+        run_name_tags = [
+            uid,
+            net_uid,
+            datetime.now().strftime("%Y_%m_%d_%H_%M_%S"),
+        ]
+
+        # Compose run name
+        run_name = "_".join(run_name_tags)
+
+        # inits wandb in case it hasn't been inited yet
+        self.wandb_run = wandb.init(
+            name=run_name,
+            project=settings.WANDB_PROJECT_NAME_MINER,
+            entity=settings.WANDB_ENTITY,
+            config=self.config,  # TODO: Check whether we really want this
+            mode="online" if settings.WANDB_ON else "offline",
+            tags=tags,
+        )
+
+    def log_event(
+        self,
+        synapse: StreamPromptingSynapse,
+        timing: float,
+        messages,
+        accumulated_chunks: list[str] = [],
+        accumulated_chunks_timings: list[float] = [],
+        extra_info: dict = {},
+    ):
+        if not getattr(self, "wandb_run", None):
+            self.init_wandb()
+
+        dendrite_uid = self.metagraph.hotkeys.index(synapse.dendrite.hotkey)
+        step_log = {
+            "epoch_time": timing,
+            # TODO: add block to logs in the future in a way that doesn't impact performance
+            # "block": self.block,
+            "messages": messages,
+            "accumulated_chunks": accumulated_chunks,
+            "accumulated_chunks_timings": accumulated_chunks_timings,
+            "validator_uid": dendrite_uid,
+            "validator_ip": synapse.dendrite.ip,
+            "validator_coldkey": self.metagraph.coldkeys[dendrite_uid],
+            "validator_hotkey": self.metagraph.hotkeys[dendrite_uid],
+            "validator_stake": self.metagraph.S[dendrite_uid].item(),
+            "validator_trust": self.metagraph.T[dendrite_uid].item(),
+            "validator_incentive": self.metagraph.I[dendrite_uid].item(),
+            "validator_consensus": self.metagraph.C[dendrite_uid].item(),
+            "validator_dividends": self.metagraph.D[dendrite_uid].item(),
+            "miner_stake": self.metagraph.S[self.uid].item(),
+            "miner_trust": self.metagraph.T[self.uid].item(),
+            "miner_incentive": self.metagraph.I[self.uid].item(),
+            "miner_consensus": self.metagraph.C[self.uid].item(),
+            "miner_dividends": self.metagraph.D[self.uid].item(),
+            **extra_info,
+        }
+
+        logger.info("Logging event to wandb...", step_log)
+        wandb.log(step_log)
+
+    def log_status(self):
+        m = self.metagraph
+        logger.info(
+            f"Miner running:: network: {self.subtensor.network} | step: {self.step} | uid: {self.uid} | trust: {m.trust[self.uid]:.3f} | emission {m.emission[self.uid]:.3f}"
+        )
