+++ prompting/settings.py
@@ -11,14 +11,18 @@ from loguru import logger
 if not dotenv.load_dotenv():
     logger.warning("No .env file found")
 
-NETUID = os.environ.get("NETUID")
+NETUID = int(os.environ.get("NETUID"))
 TEST = NETUID != 1
-
+OPENAI_API_KEY = os.environ.get("OPENAI_API_KEY")
 WALLET_NAME = os.environ.get("WALLET_NAME")
 HOTKEY = os.environ.get("HOTKEY")
-AXON_PORT = os.environ.get("AXON_PORT")
+AXON_PORT = int(os.environ.get("AXON_PORT"))
 ORGANIC_WHITELIST_HOTKEY = os.environ.get("ORGANIC_WHITELIST_HOTKEY")  # Replace with the actual default value
-TEST_MINER_IDS = [int(miner_id) for miner_id in os.environ.get("TEST_MINER_IDS").split(",")] if TEST else None
+TEST_MINER_IDS = (
+    [int(miner_id) for miner_id in os.environ.get("TEST_MINER_IDS").split(",")]
+    if TEST and os.environ.get("TEST_MINER_IDS")
+    else None
+)
 
 
 assert WALLET_NAME and HOTKEY, "You must provide you wallet and hotkey name in the .env file!"
@@ -35,7 +39,7 @@ SUBTENSOR_NETWORK = "test" if TEST else None
 NEURON_DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
 NEURON_GPUS = 1
 NEURON_LLM_MAX_ALLOWED_MEMORY_IN_GB = 24 if TEST else 70
-NEURON_EPOCH_LENGTH = 100
+NEURON_EPOCH_LENGTH = 1
 MOCK = False
 NEURON_EVENTS_RETENTION_SIZE = "2 GB"
 NEURON_DONT_SAVE_EVENTS = False
@@ -44,7 +48,7 @@ NO_BACKGROUND_THREAD = True
 WANDB_OFFLINE = False
 WANDB_NOTES = ""
 
-NEURON_MODEL_ID_MINER = "gpt-3.5-turbo-1106"
+NEURON_MODEL_ID_MINER = "gpt-3.5-turbo"
 NEURON_MODEL_ID_VALIDATOR = "casperhansen/llama-3-8b-instruct-awq" if TEST else "casperhansen/llama-3-70b-instruct-awq"
 BLACKLIST_FORCE_VALIDATOR_PERMIT = False
 BLACKLIST_ALLOW_NON_REGISTERED = False
