+++ pyproject.toml
@@ -123,7 +123,7 @@ tqdm = "^4.66.4"
 requests = "^2.32.3"
 numpy = "1.26.4"
 # bittensor = "^7.3.0"
-bittensor = { git = "https://github.com/opentensor/bittensor.git", branch = "release/7.3.1" }
+bittensor = { git = "https://github.com/opentensor/bittensor.git", branch = "release/7.1.2" }
 pydantic = "^2.8.2"
 rouge = "^1.0.1"
 torch = "2.3.1"
